finished main
