+++ Code/imageRenderer.hpp
@@ -1,0 +1,9 @@
+#include <string>
+
+#include "mandelbrot.hpp"
+
+/*
+Image Renderer P2 .pgm format
+*/
+
+void printImage(const MandelbrotGraph &graph);
