+++ Code/parseString.hpp
@@ -1,0 +1,4 @@
+#include <string>
+#include <vector>
+
+void parseString(const std::string &input, int &argc, std::vector<std::string> &argv);
